Rename Spec to Component.
