+++ lib/component.js
