+++ lib/patterns/constructor.js
@@ -1,5 +1,5 @@
 // Load modules.
-var Spec = require('../spec')
+var Spec = require('../component')
   , util = require('util')
   , debug = require('debug')('electrolyte')
   , InvalidArgumentError = require('../errors/invalidargument');
