+++ lib/patterns/literal.js
@@ -1,5 +1,5 @@
 // Load modules.
-var Spec = require('../spec')
+var Spec = require('../component')
   , util = require('util');
